chore: adds FilterSlider doc
